Adding some documentation, renamed "knowledge_graphs" to "contexts"
