+++ allennlp/data/dataset_readers/wikitables/wikitables_dataset_reader.py
@@ -18,7 +18,7 @@ from allennlp.data.instance import Instance
 from allennlp.data.token_indexers import TokenIndexer, SingleIdTokenIndexer, TokenCharactersIndexer
 from allennlp.data.tokenizers import Token, Tokenizer, WordTokenizer
 from allennlp.data.tokenizers.word_splitter import SpacyWordSplitter
-from allennlp.semparse.knowledge_graphs import TableQuestionKnowledgeGraph
+from allennlp.semparse.contexts import TableQuestionKnowledgeGraph
 from allennlp.semparse.type_declarations import wikitables_type_declaration as wt_types
 from allennlp.semparse.worlds import WikiTablesWorld
 from allennlp.semparse.worlds.world import ParsingError
