+++ allennlp/data/dataset_readers/wikitables/wikitables_preprocessed_dataset_reader.py
@@ -16,7 +16,7 @@ from allennlp.data.fields import MetadataField, ProductionRuleField, TextField
 from allennlp.data.instance import Instance
 from allennlp.data.token_indexers import TokenIndexer, SingleIdTokenIndexer, TokenCharactersIndexer
 from allennlp.data.tokenizers import Token
-from allennlp.semparse.knowledge_graphs import TableQuestionKnowledgeGraph
+from allennlp.semparse.contexts import TableQuestionKnowledgeGraph
 from allennlp.semparse.worlds import WikiTablesWorld
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
