+++ allennlp/data/fields/knowledge_graph_field.py
@@ -16,7 +16,7 @@ from allennlp.data.tokenizers.token import Token
 from allennlp.data.tokenizers.tokenizer import Tokenizer
 from allennlp.data.vocabulary import Vocabulary
 from allennlp.nn import util as nn_util
-from allennlp.semparse.knowledge_graphs.knowledge_graph import KnowledgeGraph
+from allennlp.semparse.contexts.knowledge_graph import KnowledgeGraph
 
 TokenList = List[TokenType]  # pylint: disable=invalid-name
