+++ allennlp/semparse/contexts/knowledge_graph.py
