+++ tests/data/fields/knowledge_graph_field_test.py
@@ -11,7 +11,7 @@ from allennlp.data.fields import KnowledgeGraphField
 from allennlp.data.token_indexers import SingleIdTokenIndexer, TokenCharactersIndexer
 from allennlp.data.tokenizers import WordTokenizer
 from allennlp.data.tokenizers.word_splitter import SpacyWordSplitter
-from allennlp.semparse.knowledge_graphs import TableQuestionKnowledgeGraph
+from allennlp.semparse.contexts import TableQuestionKnowledgeGraph
 
 
 class KnowledgeGraphFieldTest(AllenNlpTestCase):
