+++ tests/semparse/worlds/world_test.py
@@ -6,7 +6,7 @@ from overrides import overrides
 from allennlp.common.testing import AllenNlpTestCase
 from allennlp.data.tokenizers import Token
 from allennlp.semparse import ParsingError, World
-from allennlp.semparse.knowledge_graphs import TableQuestionKnowledgeGraph
+from allennlp.semparse.contexts import TableQuestionKnowledgeGraph
 from allennlp.semparse.worlds import NlvrWorld, WikiTablesWorld
